# admin_blog/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "思琼碧落的博客系统 API",
    "description": "个人博客系统的完整REST API文档，支持用户管理、文章发布、分类管理、音乐分享和系统配置等功能。v2.1.0更新：修复了用户角色和配置类型定义，补充了缺失的API接口。",
    "version": "2.1.0",
    "contact": {
      "name": "思琼碧落",
      "email": "zzb@siqiongbiluo.love",
      "url": "https://siqiongbiluo.love"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9000/api",
      "description": "本地开发环境"
    },
    {
      "url": "https://siqiongbiluo.love/api",
      "description": "生产环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT认证，请在请求头中添加：Authorization: Bearer {token}"
      }
    },
    "schemas": {
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "标签ID"
          },
          "name": {
            "type": "string",
            "description": "标签名称",
            "example": "Docker"
          },
          "slug": {
            "type": "string",
            "description": "标签别名（用于URL）",
            "example": "docker"
          },
          "description": {
            "type": "string",
            "description": "标签描述",
            "nullable": true,
            "example": "容器技术相关文章"
          },
          "color": {
            "type": "string",
            "description": "标签颜色（十六进制格式）",
            "default": "#007bff",
            "example": "#007bff"
          },
          "icon": {
            "type": "string",
            "description": "标签图标（CSS类名或图标名）",
            "nullable": true,
            "example": "fab fa-docker"
          },
          "article_count": {
            "type": "integer",
            "description": "使用该标签的文章数量",
            "default": 0,
            "example": 5
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "操作失败"
          },
          "error": {
            "type": "string",
            "description": "错误详情"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/articles/{id}/tags": {
      "get": {
        "tags": ["文章管理"],
        "summary": "获取文章的标签",
        "description": "获取指定文章的所有标签（公开访问）",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "文章ID",
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Tag"}
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "文章不存在",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
            }
          }
        }
      }
    },
      "post": {
        "tags": ["文章管理"],
        "summary": "为文章添加标签",
        "description": "为指定文章添加标签（需要登录，作者或管理员权限）",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "文章ID",
            "schema": {"type": "integer"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tags"],
                "properties": {
                  "tags": {
                    "oneOf": [
                      {
                        "type": "array",
                        "description": "标签名称数组",
                        "items": {"type": "string"},
                        "example": ["Docker", "部署", "博客"]
                      },
                      {
                    "type": "array",
                    "description": "标签ID数组",
                    "items": {"type": "integer"},
                    "example": [1, 2, 3]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "标签添加成功",
            "content": {
              "application/json": {
                "schema": {
                      "type": "object",
                      "properties": {
                    "success": {"type": "boolean", "example": true},
                    "message": {"type": "string", "example": "标签添加成功"},
                        "data": {
                          "type": "object",
                          "properties": {
                        "added": {"type": "integer", "description": "添加的标签数量"},
                        "existing": {"type": "integer", "description": "已存在的标签数量"},
                        "tags": {
                              "type": "array",
                          "items": {"$ref": "#/components/schemas/Tag"}
                            }
                          }
                        }
                      }
                }
              }
            }
          },
          "403": {
            "description": "无权限操作",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
            }
          }
        }
      }
    },
      "put": {
        "tags": ["文章管理"],
        "summary": "设置文章标签",
        "description": "设置文章的标签（替换所有现有标签）（需要登录，作者或管理员权限）",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "文章ID",
            "schema": {"type": "integer"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tags"],
                "properties": {
                  "tags": {
                    "oneOf": [
                      {
                        "type": "array",
                        "description": "标签名称数组",
                        "items": {"type": "string"},
                        "example": ["Vue", "前端", "教程"]
          },
          {
                        "type": "array",
                        "description": "标签ID数组",
                        "items": {"type": "integer"},
                        "example": [4, 5, 6]
          }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "标签设置成功",
            "content": {
              "application/json": {
                "schema": {
                      "type": "object",
                      "properties": {
                    "success": {"type": "boolean", "example": true},
                    "message": {"type": "string", "example": "文章标签设置成功"},
                        "data": {
                              "type": "array",
                      "items": {"$ref": "#/components/schemas/Tag"}
                            }
                          }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["文章管理"],
        "summary": "从文章移除标签",
        "description": "从指定文章移除指定标签（需要登录，作者或管理员权限）",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "文章ID",
            "schema": {"type": "integer"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tagIds"],
                "properties": {
                  "tagIds": {
                    "type": "array",
                    "items": {"type": "integer"},
                    "description": "要移除的标签ID数组",
                    "example": [1, 2]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "标签移除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "message": {"type": "string", "example": "标签移除成功"},
                    "data": {
                      "type": "object",
                      "properties": {
                        "removed": {"type": "integer", "description": "移除的标签数量"},
                        "remaining": {
                          "type": "array",
                          "items": {"$ref": "#/components/schemas/Tag"},
                          "description": "剩余的标签"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/articles/tags/{tagIds}": {
      "get": {
        "tags": ["文章管理"],
        "summary": "根据标签获取文章（GET方式）",
        "description": "根据标签ID获取文章列表，支持多种筛选条件和分页",
        "parameters": [
          {
            "name": "tagIds",
            "in": "path",
          "required": true,
            "description": "逗号分隔的标签ID列表",
            "schema": {"type": "string"},
            "example": "1,2,3"
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {"type": "integer", "minimum": 1, "default": 1}
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页数量",
            "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 10}
          },
          {
            "name": "matchType",
            "in": "query",
            "description": "标签匹配类型：any-任意匹配，all-全部匹配",
            "schema": {"type": "string", "enum": ["any", "all"], "default": "any"}
          },
          {
            "name": "category_id",
            "in": "query",
            "description": "分类ID筛选",
            "schema": {"type": "integer"}
          },
          {
            "name": "author_id",
            "in": "query",
            "description": "作者ID筛选",
            "schema": {"type": "integer"}
                },
          {
            "name": "keyword",
            "in": "query",
            "description": "搜索关键词",
            "schema": {"type": "string"}
            }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {"$ref": "#/components/schemas/PaginationResponse"},
                    {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                            "articles": {
                              "type": "array",
                              "items": {"$ref": "#/components/schemas/Article"}
                            }
                          }
                      }
                    }
                  }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/tags": {
      "get": {
        "tags": ["标签管理"],
        "summary": "获取标签列表",
        "description": "获取所有标签或分页获取标签列表",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码（可选，不提供则返回所有标签）",
            "schema": {"type": "integer", "minimum": 1, "default": 1}
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页数量（可选）",
            "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 10}
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "排序字段",
            "schema": {"type": "string", "enum": ["article_count", "created_at", "name"], "default": "article_count"}
          },
          {
            "name": "order",
            "in": "query",
            "description": "排序方向",
            "schema": {"type": "string", "enum": ["ASC", "DESC"], "default": "DESC"}
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {
                      "type": "object",
                      "properties": {
                        "tags": {
                          "type": "array",
                          "items": {"$ref": "#/components/schemas/Tag"}
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["标签管理"],
        "summary": "创建标签",
        "description": "创建新标签（需要管理员权限）",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/TagInput"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "message": {"type": "string", "example": "标签创建成功"},
                    "data": {"$ref": "#/components/schemas/Tag"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
...
    "/tags/slug/{slug}": {
      "get": {
        "tags": ["标签管理"],
        "summary": "根据别名获取标签详情",
        "description": "根据标签别名获取标签详细信息（公开访问）",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "标签别名",
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {"$ref": "#/components/schemas/Tag"}
                  }
                }
              }
            }
          },
          "404": {
            "description": "标签不存在",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
...
    "/tags/{id}/articles": {
      "get": {
        "tags": ["标签管理"],
        "summary": "获取标签的文章列表",
        "description": "获取使用指定标签的文章列表",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "标签ID",
            "schema": {"type": "integer"}
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {"type": "integer", "minimum": 1, "default": 1}
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页数量",
            "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 10}
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {
                      "type": "object",
                      "properties": {
                        "tag": {"$ref": "#/components/schemas/Tag"},
                        "articles": {
                          "type": "array",
                          "items": {"$ref": "#/components/schemas/Article"}
                        },
                        "pagination": {"$ref": "#/components/schemas/Pagination"}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
} 